requirements.txt
numpy~=2.0.2
typing_extensions~=4.14.0
pillow~=11.2.1
pip~=23.2.1
wheel~=0.45.1
Jinja2~=3.1.6
gast~=0.6.0
h5py~=3.14.0
pytz~=2025.2
rich~=14.0.0
Pygments~=2.19.1
requests~=2.32.4
keras~=3.10.0
colorama~=0.4.6
ml_dtypes~=0.5.1
tensorflow~=2.19.0
optree~=0.16.0
packaging~=25.0
scipy~=1.15.3
torch~=2.7.1
pandas~=2.3.0
namex~=0.1.0
matplotlib~=3.10.3
mdurl~=0.1.2
llvmlite~=0.44.0
numba~=0.61.2
setuptools~=80.9.0
six~=1.17.0
mpmath~=1.3.0
sympy~=1.14.0
astunparse~=1.6.3
tensorboard~=2.19.0
filelock~=3.18.0
tqdm~=4.67.1
torchvision~=0.22.1
opt_einsum~=3.4.0
networkx~=3.5
urllib3~=2.4.0
fsspec~=2025.5.1
MarkupSafe~=3.0.2
python-dateutil~=2.9.0.post0
certifi~=2025.4.26
natsort~=8.4.0
roifile~=2025.5.10
tifffile~=2025.6.11
fastremap~=1.16.1
cellpose~=3.1.1
imagecodecs~=2025.3.30
Markdown~=3.8
idna~=3.10
contourpy~=1.3.2
fonttools~=4.58.4
pyparsing~=3.2.3
termcolor~=3.1.0
cycler~=0.12.1
kiwisolver~=1.4.8
flatbuffers~=25.2.10
wrapt~=1.17.2
Werkzeug~=3.1.3